+++ orchestrator/CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+#### Added
+
+* Added support for reading Azure Storage credentials from the K8s secret ([#27](https://github.com/stjude-rust-labs/planetary/pull/27)).
+
+#### Fixed
+
+* Fixed returning a 403 for when there is an invalid or missing authorization
+  header ([#26](https://github.com/stjude-rust-labs/planetary/pull/26)).
+* Fixed incorrect task state when a system error was encountered ([#28](https://github.com/stjude-rust-labs/planetary/pull/28)).
+
 ## v0.1.0 (2025-10-13)
 
 #### Added